fixed big int

+++ src/rust/src/bindings/dom.rs
@@ -96,7 +96,7 @@ fn remove_change_event_handler(callback_id: &Rc<ExternRef>) {
 }
 
 #[no_mangle]
-pub extern "C" fn web_handle_change_event(callback_id: u64, allocation_id: u32) {
+pub extern "C" fn web_handle_change_event(callback_id: u32, allocation_id: u32) {
     ELEMENT_CHANGE_HANDLERS.with(|s| {
 
         let handler = s.lock().map(|mut s| {
@@ -117,7 +117,7 @@ pub fn add_change_event_listener(element: &ExternRef, handler: impl FnMut(Change
                 const text = e.target.value;
                 const buffer = (new TextEncoder()).encode(text);
                 const allocationId = writeBufferToMemory(buffer);
-                wasmModule.instance.exports.web_handle_change_event(BigInt(objectId), allocationId);
+                wasmModule.instance.exports.web_handle_change_event(objectId, allocationId);
             };
             const objectId = storeObject(handler);
             element.addEventListener("change",handler);
@@ -170,7 +170,7 @@ thread_local! {
 }
 
 #[no_mangle]
-pub extern "C" fn web_handle_mouse_event_handler(callback_id: u64, x: f64, y: f64) {
+pub extern "C" fn web_handle_mouse_event_handler(callback_id: u32, x: f64, y: f64) {
 
     MOUSE_EVENT_HANDLER.with(|s| {
         s.call(callback_id as u32, MouseEvent { offset_x: x, offset_y: y });
@@ -182,7 +182,7 @@ pub fn element_add_click_listener(element: &ExternRef, handler: impl FnMut(Mouse
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_mouse_event_handler(BigInt(objectId),e.offsetX, e.offsetY);
+                wasmModule.instance.exports.web_handle_mouse_event_handler(objectId,e.offsetX, e.offsetY);
             };
             const objectId = storeObject(handler);
             element.addEventListener("click",handler);
@@ -209,7 +209,7 @@ pub fn element_add_mouse_move_listener(element: &ExternRef, handler: impl FnMut(
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_mouse_event_handler(BigInt(objectId),e.offsetX, e.offsetY);
+                wasmModule.instance.exports.web_handle_mouse_event_handler(objectId,e.offsetX, e.offsetY);
             };
             const objectId = storeObject(handler);
             element.addEventListener("mousemove",handler);
@@ -235,7 +235,7 @@ pub fn element_add_mouse_down_listener(element: &ExternRef, handler: impl FnMut(
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_mouse_event_handler(BigInt(objectId),e.offsetX, e.offsetY);
+                wasmModule.instance.exports.web_handle_mouse_event_handler(objectId,e.offsetX, e.offsetY);
             };
             const objectId = storeObject(handler);
             element.addEventListener("mousedown",handler);
@@ -261,7 +261,7 @@ pub fn element_add_mouse_up_listener(element: &ExternRef, handler: impl FnMut(Mo
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_mouse_event_handler(BigInt(objectId),e.offsetX, e.offsetY);
+                wasmModule.instance.exports.web_handle_mouse_event_handler(objectId,e.offsetX, e.offsetY);
             };
             const objectId = storeObject(handler);
             element.addEventListener("mouseup",handler);
@@ -305,7 +305,7 @@ fn remove_keyboard_event_handler(function_handle: &Rc<ExternRef>) {
 }
 
 #[no_mangle]
-pub extern "C" fn web_handle_keyboard_event_handler(callback_id: u64, key_code: f64) {
+pub extern "C" fn web_handle_keyboard_event_handler(callback_id: u32, key_code: f64) {
 
     KEYBOARD_EVENT_HANDLERS.with(|s| {
 
@@ -323,7 +323,7 @@ pub fn element_add_key_down_listener(element: &ExternRef, handler: impl FnMut(Ke
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_keyboard_event_handler(BigInt(objectId),e.keyCode);
+                wasmModule.instance.exports.web_handle_keyboard_event_handler(objectId,e.keyCode);
             };
             const objectId = storeObject(handler);
             element.addEventListener("keydown",handler);
@@ -345,7 +345,7 @@ pub fn element_add_key_up_listener(element: &ExternRef, handler: impl FnMut(Keyb
     let code = r#"
         function(element ){
             const handler = (e) => {
-                wasmModule.instance.exports.web_handle_keyboard_event_handler(BigInt(objectId),e.keyCode);
+                wasmModule.instance.exports.web_handle_keyboard_event_handler(objectId,e.keyCode);
             };
             const objectId = storeObject(handler);
             element.addEventListener("keyup",handler);
